+++ Backend/src/modules/emails/email.controller.ts
@@ -1,5 +1,14 @@
 import { NextFunction, Request, Response } from 'express';
-import { ArchiveEmailBody, DeleteEmailSchema, GetAllEmailsSchema, SearchEmailBody, StarEmailBody, UnreadEmailBody } from './email.schema.js';
+import {
+    ArchiveEmailBody,
+    ComposeEmailBody,
+    DeleteEmailSchema,
+    GetAllEmailsSchema,
+    SearchEmailBody,
+    SearchOtherContactsBody,
+    StarEmailBody,
+    UnreadEmailBody,
+} from './email.schema.js';
 import { EmailService } from './email.service.js';
 import { GetAllEmailsFilters } from './email.types.js';
 
@@ -126,6 +135,37 @@ export class EmailController {
                 throw new Error('Email ID is required');
             }
             const email = await this.emailService.unreadEmails(emailIds, Boolean(unread));
+            res.send(email);
+        } catch (error) {
+            next(error);
+        }
+    };
+
+    public composeEmail = async (req: Request<object, object, ComposeEmailBody, object>, res: Response, next: NextFunction): Promise<void> => {
+        try {
+            const { accountId, to, subject, body } = req.body;
+            if (!accountId || !to || !subject || !body) {
+                throw new Error('Account ID, To, subject and body are required');
+            }
+            const email = await this.emailService.composeEmail(req.body);
+            res.send(email);
+        } catch (error) {
+            next(error);
+        }
+    };
+
+    public searchOtherContacts = async (
+        req: Request<object, object, SearchOtherContactsBody, object>,
+        res: Response,
+        next: NextFunction,
+    ): Promise<void> => {
+        try {
+            const { searchText } = req.body;
+            const userId = req.user?.id;
+            if (!userId) {
+                throw new Error('User ID is required');
+            }
+            const email = await this.emailService.searchOtherContacts(userId, searchText);
             res.send(email);
         } catch (error) {
             next(error);