+++ Backend/src/providers/outlook/outlook.service.ts
@@ -16,6 +16,9 @@ import {
     OutlookUserProfile,
 } from './outlook.types.js';
 import * as OutlookUtils from './outlook.utils.js';
+import { ComposeEmailBody } from '@modules/emails/email.schema.js';
+import { compressString } from '@utils/compression.js';
+import { SearchOtherContactsResponse } from '@modules/emails/email.types.js';
 
 export class OutlookService {
     private outlookApi: OutlookApi;
@@ -158,7 +161,7 @@ export class OutlookService {
         }
     }
 
-    private async parseEmailDetailsIntoPlainObject(accountId: string, email: OutlookMessageObjectFull): Promise<EmailInput> {
+    private parseEmailDetailsIntoPlainObject(accountId: string, email: OutlookMessageObjectFull): EmailInput {
         try {
             const emailObject: EmailInput = {
                 accountId,
@@ -170,8 +173,8 @@ export class OutlookService {
                 bcc: email.bccRecipients.map((val) => val.emailAddress.address),
                 subject: email.subject,
                 body: email.bodyPreview || '',
-                bodyPlain: email.body.content,
-                bodyHtml: email.body.content,
+                bodyPlain: compressString(email.body.content),
+                bodyHtml: compressString(email.body.content),
                 receivedAt: new Date(email.receivedDateTime),
                 isRead: email.isRead,
                 folders: [email.parentFolderId],
@@ -337,6 +340,40 @@ export class OutlookService {
         } catch (err) {
             const errorMessage = err instanceof Error ? err.message : String(err);
             logger.error(`Error in OutlookService.deleteFolder: ${errorMessage}`, { error: err });
+            throw err;
+        }
+    }
+
+    async sendMail(composeEmailData: ComposeEmailBody): Promise<OutlookMessageObjectFull> {
+        try {
+            const { accountId, to, subject, body } = composeEmailData;
+            const messageBody = OutlookUtils.buildOutlookMessagePayload(to, subject, body, 'HTML');
+            const response = await OutlookApi.createDraftMessage(accountId, messageBody);
+            await OutlookApi.sendDraftMessage(accountId, response.id);
+            const emailDetails = await OutlookApi.getMessageDetails(accountId, response.id);
+            const emailData = this.parseEmailDetailsIntoPlainObject(accountId, emailDetails);
+            await EmailRepository.upsertEmailsInBulk([emailData]);
+            return response;
+        } catch (err) {
+            const errorMessage = err instanceof Error ? err.message : String(err);
+            logger.error(`Error in OutlookService.sendMail: ${errorMessage}`, { error: err });
+            throw err;
+        }
+    }
+
+    async searchContacts(accountId: string, searchText: string): Promise<SearchOtherContactsResponse[]> {
+        try {
+            const response = await OutlookApi.searchContacts(accountId, searchText);
+            const contacts = response.value
+                .filter((contact) => contact.scoredEmailAddresses && contact.scoredEmailAddresses.length > 0)
+                .map((contact) => ({
+                    email: contact.scoredEmailAddresses[0].address,
+                    name: contact.displayName,
+                }));
+            return contacts;
+        } catch (err) {
+            const errorMessage = err instanceof Error ? err.message : String(err);
+            logger.error(`Error in OutlookService.searchContacts: ${errorMessage}`, { error: err });
             throw err;
         }
     }