+++ Frontend/src/modules/folders/services/useFolderApi.ts
@@ -1,11 +1,11 @@
 import { FOLDER_KEYS } from '@shared/config/query-keys';
-import { UpdateAPIResponse } from '@shared/types/api.types';
-import { CreateFolderBodyParams, FolderAttributes, GetALlFolderResponse, GetAllFoldersRequestOptions } from '@shared/types/folder.types';
+import { PaginatedDataResponse, UpdateAPIResponse } from '@shared/types/api.types';
+import { CreateFolderBodyParams, FolderAttributes, GetAllFoldersRequestOptions } from '@shared/types/folder.types';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { createFolder, deleteFolder, getAllFolders, getFolderDetails, updateFolder } from './folder.api';
 
 export const useGetAllFolders = (options: GetAllFoldersRequestOptions | null) => {
-    return useQuery<GetALlFolderResponse, Error>({
+    return useQuery<PaginatedDataResponse<FolderAttributes>, Error>({
         queryKey: [FOLDER_KEYS.FOLDERS, options?.page, options?.size, options?.filters],
         queryFn: () => getAllFolders(options!),
         enabled: !!options?.userId,